adicionando os triggers relacionados a Local e suas especializações

+++ BD/restricao_integridade.sql
@@ -144,7 +144,9 @@ EXECUTE FUNCTION verifica_treinador_lider();
 CREATE OR REPLACE FUNCTION deleta_treinador_pc()
 RETURNS TRIGGER AS $$
 BEGIN
-    DELETE FROM treinador WHERE treinador_id = OLD.treinador_id;
+    IF OLD.treinador_id IN (SELECT treinador_id FROM PC) THEN
+        DELETE FROM treinador WHERE treinador_id = OLD.treinador_id;
+    END IF;
     RETURN OLD;
 END;
 $$ LANGUAGE plpgsql;
@@ -158,7 +160,9 @@ EXECUTE FUNCTION deleta_treinador_pc();
 CREATE OR REPLACE FUNCTION deleta_treinador_npc()
 RETURNS TRIGGER AS $$
 BEGIN
-    DELETE FROM treinador WHERE treinador_id = OLD.treinador_id;
+    IF OLD.treinador_id IN (SELECT treinador_id FROM NPC) THEN
+        DELETE FROM treinador WHERE treinador_id = OLD.treinador_id;
+    END IF;
     RETURN OLD;
 END;
 $$ LANGUAGE plpgsql;
@@ -172,7 +176,9 @@ EXECUTE FUNCTION deleta_treinador_npc();
 CREATE OR REPLACE FUNCTION deleta_treinador_lider()
 RETURNS TRIGGER AS $$
 BEGIN
-    DELETE FROM treinador WHERE treinador_id = OLD.treinador_id;
+    IF OLD.treinador_id IN (SELECT treinador_id FROM lider) THEN
+        DELETE FROM treinador WHERE treinador_id = OLD.treinador_id;
+    END IF;
     RETURN OLD;
 END;
 $$ LANGUAGE plpgsql;
@@ -202,6 +208,123 @@ CREATE TRIGGER delete_treinador
 BEFORE DELETE ON treinador
 FOR EACH ROW
 EXECUTE FUNCTION deleta_especializacao_treinador();
+
+------------------------------------------------------------
+
+CREATE OR REPLACE FUNCTION verifica_local_pokecenter() 
+RETURNS TRIGGER AS $$
+BEGIN
+    IF NEW.local_id IN (SELECT local_id FROM Ginasio) THEN
+        RAISE EXCEPTION 'Local já existe como Ginásio';
+    ELSIF NEW.local_id IN (SELECT local_id FROM Pokemart) THEN
+        RAISE EXCEPTION 'Local já existe como Pokemart';
+    ELSIF NEW.local_id IN (SELECT local_id FROM Caminho) THEN
+        RAISE EXCEPTION 'Local já existe como Caminho';
+    ELSIF NEW.local_id IN (SELECT local_id FROM Zona_de_captura) THEN
+        RAISE EXCEPTION 'Local já existe como Zona de Captura';
+    ELSIF NEW.local_id NOT IN (SELECT local_id FROM Local_) THEN
+        RAISE EXCEPTION 'Local precisa primeiro ser adicionado a tabela Local_';
+    END IF;
+    RETURN NEW;
+END;
+$$ LANGUAGE plpgsql;
+
+CREATE TRIGGER insercao_pokecenter
+BEFORE INSERT ON Pokecenter
+FOR EACH ROW
+EXECUTE FUNCTION verifica_local_pokecenter();
+
+CREATE OR REPLACE FUNCTION verifica_local_ginasio() 
+RETURNS TRIGGER AS $$
+BEGIN
+    IF NEW.local_id IN (SELECT local_id FROM Pokecenter) THEN
+        RAISE EXCEPTION 'Local já existe como Pokecenter';
+    ELSIF NEW.local_id IN (SELECT local_id FROM Pokemart) THEN
+        RAISE EXCEPTION 'Local já existe como Pokemart';
+    ELSIF NEW.local_id IN (SELECT local_id FROM Caminho) THEN
+        RAISE EXCEPTION 'Local já existe como Caminho';
+    ELSIF NEW.local_id IN (SELECT local_id FROM Zona_de_captura) THEN
+        RAISE EXCEPTION 'Local já existe como Zona de Captura';
+    ELSIF NEW.local_id NOT IN (SELECT local_id FROM Local_) THEN
+        RAISE EXCEPTION 'Local precisa primeiro ser adicionado a tabela Local_';
+    END IF;
+    RETURN NEW;
+END;
+$$ LANGUAGE plpgsql;
+
+CREATE TRIGGER insercao_ginasio
+BEFORE INSERT ON Ginasio
+FOR EACH ROW
+EXECUTE FUNCTION verifica_local_ginasio();
+
+CREATE OR REPLACE FUNCTION verifica_local_pokemart() 
+RETURNS TRIGGER AS $$
+BEGIN
+    IF NEW.local_id IN (SELECT local_id FROM Pokecenter) THEN
+        RAISE EXCEPTION 'Local já existe como Pokecenter';
+    ELSIF NEW.local_id IN (SELECT local_id FROM Ginasio) THEN
+        RAISE EXCEPTION 'Local já existe como Ginásio';
+    ELSIF NEW.local_id IN (SELECT local_id FROM Caminho) THEN
+        RAISE EXCEPTION 'Local já existe como Caminho';
+    ELSIF NEW.local_id IN (SELECT local_id FROM Zona_de_captura) THEN
+        RAISE EXCEPTION 'Local já existe como Zona de Captura';
+    ELSIF NEW.local_id NOT IN (SELECT local_id FROM Local_) THEN
+        RAISE EXCEPTION 'Local precisa primeiro ser adicionado a tabela Local_';
+    END IF;
+    RETURN NEW;
+END;
+$$ LANGUAGE plpgsql;
+
+CREATE TRIGGER insercao_pokemart
+BEFORE INSERT ON Pokemart
+FOR EACH ROW
+EXECUTE FUNCTION verifica_local_pokemart();
+
+CREATE OR REPLACE FUNCTION verifica_local_caminho() 
+RETURNS TRIGGER AS $$
+BEGIN
+    IF NEW.local_id IN (SELECT local_id FROM Pokecenter) THEN
+        RAISE EXCEPTION 'Local já existe como Pokecenter';
+    ELSIF NEW.local_id IN (SELECT local_id FROM Ginasio) THEN
+        RAISE EXCEPTION 'Local já existe como Ginásio';
+    ELSIF NEW.local_id IN (SELECT local_id FROM Pokemart) THEN
+        RAISE EXCEPTION 'Local já existe como Pokemart';
+    ELSIF NEW.local_id IN (SELECT local_id FROM Zona_de_captura) THEN
+        RAISE EXCEPTION 'Local já existe como Zona de Captura';
+    ELSIF NEW.local_id NOT IN (SELECT local_id FROM Local_) THEN
+        RAISE EXCEPTION 'Local precisa primeiro ser adicionado a tabela Local_';
+    END IF;
+    RETURN NEW;
+END;
+$$ LANGUAGE plpgsql;
+
+CREATE TRIGGER insercao_caminho
+BEFORE INSERT ON Caminho
+FOR EACH ROW
+EXECUTE FUNCTION verifica_local_caminho();
+
+CREATE OR REPLACE FUNCTION verifica_local_zona_captura() 
+RETURNS TRIGGER AS $$
+BEGIN
+    IF NEW.local_id IN (SELECT local_id FROM Pokecenter) THEN
+        RAISE EXCEPTION 'Local já existe como Pokecenter';
+    ELSIF NEW.local_id IN (SELECT local_id FROM Ginasio) THEN
+        RAISE EXCEPTION 'Local já existe como Ginásio';
+    ELSIF NEW.local_id IN (SELECT local_id FROM Pokemart) THEN
+        RAISE EXCEPTION 'Local já existe como Pokemart';
+    ELSIF NEW.local_id IN (SELECT local_id FROM Caminho) THEN
+        RAISE EXCEPTION 'Local já existe como Caminho';
+    ELSIF NEW.local_id NOT IN (SELECT local_id FROM Local_) THEN
+        RAISE EXCEPTION 'Local precisa primeiro ser adicionado a tabela Local_';
+    END IF;
+    RETURN NEW;
+END;
+$$ LANGUAGE plpgsql;
+
+CREATE TRIGGER insercao_zona_captura
+BEFORE INSERT ON Zona_de_captura
+FOR EACH ROW
+EXECUTE FUNCTION verifica_local_zona_captura();
